Add files via upload
This MachineTrader flow will allow you to download orders (open or closed) from Alpaca to a csv file on your local device.  Edit the "get Orders" function node to select either open or closed orders and to set the number of orders to download (the default is set to 10). Alpaca limit the maximum number of orders in the download to 500. To download the file, append the url string in the /api/download_orders to your instance name: yourinstance,machinetrader.io/api/download_orders and the file will download.