Add files via upload
This file contains flows that may be required to optimize your instance. The first flow reduces the size of the sqlite table "balances" each night so that no more than one week of data is stores. It updates the balances table in both the live_account and paper_account databases.