Mean Reversion Calculations for a Tickers
This flow calculates a moving average value for a ticker based on a set look back (the default is the past 100 trades). The second sentence retrieves the last trade. Subtracting the moving average from the last trade price provides a delta between the two values which can be a useful buy signal assuming prices have a tendency to return to the mean. The flow can be set to run every minute with a starting and ending time, optionally using a fixed loop to iterate 60 times a minute - once per second. Or it could be set to run every second (in the case of cryptos).