Short Top Stocks
MachineTrader continuous updates a list of the 100 top performing stocks through the day in its data center. The strategy assumes that there has been too much buying momentum and shorts the top 30 stocks 15 mins prior to market close (3:45 pm).

The stocks are held overnight and then a sequence of limit trades begins are 4 am (when extended hours begin). Each minute the flow looks to see the current profit on each ticker held.  If the target profit is reached, a limit buy order is entered. [The profit target is set in line 10 of the node "define values required for adding trade to portfolio" in Flow 1]. Any short positions  that have not been covered  the market open, are sold as market orders at  9:31 am.