Flows for Managing Limit Orders
Tickers prices are always fluctuating and -- as a result -- oftentimes limit orders are not filled if the price is moving in the wrong direction. This flow displays the limit_prices on open orders and compares it to the current price.  The trade order flow can be used to change the limit prices dynamically according to simple rules.