Trade every second
This flow allows for intraday trading second by second. The default is trading TQQQ every minute throughout the trading day (10 am - 4 pm) using market orders. The buy side buys 100 shares every time the position is empty. The sell side liquidates the positions whenever a profit of $1 is made. A trigger based on the moving average price could be added to improve the pricing on the buy side.