brief flow that updates positions table
Manual trades are entered without defining a strategy. As a result they will display in the list of positions with empty descriptions.  This flow (when added to the tab "Update on Intervals") runs every minute and updates the portfolio name as "Manual Trades" and the iD as "999999". I added it to the latest version in GitHub.