Reports performance for Interday Strategies.
This flow updates the Algos Display in the front end from the algos table.
It only works properly for "interday" strategies which are purchased and sold once. 
It relies on values supplied by Alpaca for todays gains and gains all time for each ticker that is currently held.
The function node "get totals from positions" totals  the unrealized_intraday_pl and unrealized_pl for each ticker contained in a strategy.
The percent changes are calculated from the total of the market value for each ticker.
The gains or losses are compared with the performance of the benchmark over time.