Update Intraday Algo Tracking
This flow calculates the performance of Intraday stategies which are more complicated since there can be many trades over time.
This method uses "orders_all" table which should conain a record of all "closed" orders processed by Alpaca. 
Because Alpaca only allows the downloading of the 500 most recent orders, it is necessary to download the orders daily.
Note: if you are using more than 500 closed orders a day, you will need increase the frequency of downloads.

The method we use is to sum all of the "buy" and "sell" orders in the orders_old table for each ticker 
associated with the strategy. Note that the orders_all table includes all orders ever completed. So summing the
buys and sells will include include all transactations for that ticker. Since you may want to exclude
very old transactions from the algo reporting, there is a date filter included in the sql query in function node
"get sum buys orders". The query only includes orders made after the "created_at" date.

We will also need to record if there is a current position for that
ticker. The buy and sell orders are totalled, along with the current position (if any). 

To store the values we need a new table: "intraday_buy_sell" which is created in one of the flows.
Values are stored each time the flow runs, so it starts with a "truncate" request to get a fresh table.
Totals buys, sells, and current positions are stored in this table.
Gains/loss are calculated and stored in the profit field.

The performance of the algo is determined by summing all of the profits for each ticker in the strategy.

The percent changes are calculated from the total of the market value for each ticker.
The gains or losses are compared with the performance of the benchmark over time.

Because is is sometimes useful to look at the entire orders table, we have included the api endpoint: 
[your instance].machinetrader.io/api/download_orders_old which downloads the entire orders table as a csv file.