latest version
Research Ticker

This tab is useful for determining trading signals for intraday trading strategies.  


Set the ticker you wish to research in the starting inject node, storing the ticker as a flow.ticker for future use.  



Flow 1 allows you to specify the number of trades to be analyzed in the first function  node. 

The Polygon node returns an array with the values of the last 1,000 trades. The array is passed to a Calculator Node that calculates the average price of the 1,000 trades. The Polygon node is accessed once again, this time requested on the most recent trade. The average price, the current price, and the difference between the prices are stored as flow variables. The spread between the current price and the average price is a common trigger for use in executing trading decisions.

Flow 2 is designed to statistics on a larger group of trades, in this case the past 10,000 trades. Calculations are made for the maximum, minimum, and average price. Values are stored as flow variables.
  


Flow 3 displays the results of the calculations in the debug panel.



It would be useful to display the standard deviation of the trades, although the Calculator node lacks that calculator. However, we can download an array of tickers to a csv file on our local machine using Flow 4. 


The files downloaded by entering the url in the browser:

[your instance name].machinetrader.io/api/download_output
https://pta14.machinetrader.io/api/download_output


In Chrome the file appears in the button left corner. Click and it opens in Excel if that is your default csv file reader. Create a new column of data with the change in prices. Use the array of changes with the Excel function.

 

The std dev is often a useful trading trigger and it will occur approximately 15 percent of the time. So a trading algo written to execute every minute should find a trade approximately every 15 minutes.