latest version of Update orders table flow
Instances created before 2/14/23 will need to update the "Update Orders" flow 

This flow replaces the “Update Orders” flow at the bottom of the “Update on Intervals” tab.  It allows new orders to be downloaded from Alpaca every minute and added for the orders table. It also updates the orders table with the portfolio names for each strategy so that all tickers can be associated with a strategy that is defined in the tickers_paper and ticker_live tables.

If your instance was created before 2/14/23, you may be using an older version of the orders table.  You will need to run the flows shown below.

Documentation: https://docs.google.com/document/d/1-0Y0Xdoxn74KSzsJ7QMC9msUX1r_QTVRttUg9mig5Gg/edit?usp=sharing